Drastically faster templating approach
The current templating approach is really inefficient, as it recursively loops thru every 
passed available variable for substituion and applies a regex for each one. 

It's much faster to initially parse the template and then only process those strings which are 
actually present in the template.

This revised implementation both caches the parsed templates and then only processes the string which are present.

The performance improvement is quite drastic, especially with large tables. A large table was previously taking 15s to render 
in Chrome, now takes less than 1s with this approach